backend/requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==44.0.3
defusedxml==0.7.1
Django==5.2.1
django-cors-headers==4.7.0
django-filter==25.1
django-templated-mail==1.1.1
djangorestframework==3.16.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-nested-routers==0.94.1
idna==3.10
iniconfig==2.1.0
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
oauthlib==3.2.2
packaging==25.0
Pillow==9.0.0
pluggy==1.0.0.dev0
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.22
PyJWT==2.10.1
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==4.0.0
social-auth-core==4.6.1
sqlparse==0.5.3
toml==0.10.2
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
webcolors==1.11.1
